+++ mnist/train.log
@@ -19,3 +19,22 @@
 2023-05-04 12:04:52,391 Training completed. Best loss is 0.20899625825248658, in epoch 1
 2023-05-04 12:04:54,299 Test Without Attack, Accuracy: 97.48%
 2023-05-04 12:04:56,664 Test With Attack, Method: fsgm, Epsilon: 0.1, Accuracy: 47.03%
+2023-05-04 18:28:19,743 Epoch [1/30], Loss: 0.1781
+2023-05-04 18:28:39,325 Epoch [2/30], Loss: 0.0593
+2023-05-04 18:28:58,485 Epoch [3/30], Loss: 0.0435
+2023-05-04 18:29:17,633 Epoch [4/30], Loss: 0.0350
+2023-05-04 18:29:36,745 Epoch [5/30], Loss: 0.0299
+2023-05-04 18:29:55,792 Epoch [6/30], Loss: 0.0254
+2023-05-04 18:30:14,913 Epoch [7/30], Loss: 0.0215
+2023-05-04 18:30:34,031 Epoch [8/30], Loss: 0.0196
+2023-05-04 18:30:53,030 Epoch [9/30], Loss: 0.0164
+2023-05-04 18:31:12,050 Epoch [10/30], Loss: 0.0151
+2023-05-04 18:31:31,177 Epoch [11/30], Loss: 0.0135
+2023-05-04 18:31:50,350 Epoch [12/30], Loss: 0.0135
+2023-05-04 18:32:09,982 Epoch [13/30], Loss: 0.0110
+2023-05-04 18:32:29,113 Epoch [14/30], Loss: 0.0125
+2023-05-04 18:32:48,295 Epoch [15/30], Loss: 0.0094
+2023-05-04 18:33:08,623 Epoch [16/30], Loss: 0.0093
+2023-05-04 18:33:27,702 Epoch [17/30], Loss: 0.0104
+2023-05-04 18:33:46,769 Epoch [18/30], Loss: 0.0071
+2023-05-04 18:34:06,707 Epoch [19/30], Loss: 0.0088
